// src/types/wishList.ts
import type { Product } from "@/types/product";

export interface WishlistItem {
  _id?: string;
  product?: Product;
}

export interface Root {
  status: string;
  message: string;
  data: Array<WishlistItem | Product>;
}

export interface WishListRoot {
  status: string
  count: number
  data: Daum[]
}

export interface Daum {
  sold?: number
  images?: string[]
  subcategory?: Subcategory[]
  ratingsQuantity?: number
  _id?: string
  product?: Product
  title?: string
  slug?: string
  description?: string
  quantity?: number
  price?: number
  imageCover?: string
  category?: Category
  brand?: Brand
  ratingsAverage?: number
  createdAt?: string
  updatedAt?: string
  __v?: number
  id?: string
}

export interface Subcategory {
  _id: string
  name: string
  slug: string
  category: string
}

export interface Category {
  _id: string
  name: string
  slug: string
  image: string
}

export interface Brand {
  _id: string
  name: string
  slug: string
  image: string
}